Reduce warnings on windows.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@28201 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/databases/Image/vtkStimulateReader.C
@@ -143,7 +143,7 @@ bool vtkStimulateReader::GetFilenames(const char *one_file, char *spr_name,
 {
   if (one_file == NULL)
     return false;
-  int len = strlen(one_file);
+  size_t len = strlen(one_file);
   if (len < 4)
     return false;
